Refactor theme handling in GenericButton and DarkModeToggle components to use updated DarkThemeContext and improve prop management
