+++ src/components/ToggleButton/ToggleButton.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import clsx from 'clsx';
 
-type ToggleButtonProps = {
+export type ToggleButtonProps = {
   value: boolean;
   setValue: (value: any) => void;
   size?: number;
@@ -9,7 +9,7 @@ type ToggleButtonProps = {
   style?: any;
 };
 
-export default function ToggleButton(props: ToggleButtonProps) {
+export function ToggleButton(props: ToggleButtonProps) {
   const { value, setValue, size = 32, className = '', style } = props;
 
   const isChecked = value;
